+++ lib/app.js
@@ -27,6 +27,7 @@ app.use(express.json());
 app.use(express.urlencoded());
 app.use(express.methodOverride());
 app.use(app.router);
+app.use(express.bodyParser());
 app.use(expressValidator());
 
 app.configure('dev', function() {
