+++ lib/jobs/routes/jobs.js
@@ -57,10 +57,17 @@ module.exports = function(Jobs) {
   routes.createFormValidation = function(req, res, next) {
     // create form validations are here
     req.assert('name', '1 to 30 charachors required').notEmpty().len(1, 30);
-    if (req.validationErrors() && req.validationErrors().length) {
+    var fileError;
+    var uploadFile = req.files;
+    if (JSON.stringify(uploadFile) =="{}") {
+      fileError = { param: 'configFile', msg: 'null file', value: '' };
+    }else if(!uploadFile.configFile.originalFilename.match(/\.json$/)){
+      fileError = { param: 'configFile', msg: 'only accept .js file', value: 'uploadFile.configFile.originalFilename' };
+    }
+    if (fileError || req.validationErrors() && req.validationErrors().length) {
       return res.format({
         html: function() {
-          res.render('new-job', {title: 'Create A new Viff Job', errors: req.validationErrors() });
+          res.render('new-job', {title: 'Create A new Viff Job', errors: (req.validationErrors()).push(fileError) });
         },
         json: function() {
           res.send(400, req.validationErrors());
