+++ test/helper.js
@@ -1,0 +1,44 @@
+var fs = require('fs');
+var path = require('path');
+var mongoskin = require('mongoskin');
+var db = mongoskin.db('mongodb://localhost:27017/viffService?auto_reconnect', {safe: true});
+
+var existingFile;
+var uploadsPath = path.join(__dirname, "../uploads");
+
+// recode existing files in uploads/configFile.json 
+before(function (done) {
+  fs.readdir(uploadsPath, function (err, fileList) {
+    existingFile = fileList;
+    done();
+  });
+});
+
+// remove created test uploaded files and revert db changes
+after(function (done) {
+  //remove uploaded files
+  fs.readdir(uploadsPath, function (err, fileList) {
+    fileList.forEach(function (file) {
+      if (existingFile.indexOf(file) == -1) {
+        fs.unlink(uploadsPath + '/' + file, function (err) {
+          if (err) {
+            throw err;
+          }
+        });
+      }
+    });
+    //remove db changes
+    db.collection('jobs').remove(
+      {$or: [
+        {name: "test job"},
+        {name: "db save job test"}
+        // {name: "job with id"}
+      ]},
+      function (err) {
+        if (err) {
+          throw err;
+        }
+        done();
+      });
+  });
+});
