+++ test/lib/jobs/configFile.haha
@@ -1,0 +1,3 @@
+{
+  "test": true
+}
