+++ test/lib/jobs/job.js
@@ -1,6 +1,5 @@
 var request = require('request');
 var fs = require('fs');
-var path = require('path');
 var mongoskin = require('mongoskin');
 var db = mongoskin.db('mongodb://localhost:27017/viffService?auto_reconnect', {safe: true});
 var JobsModule = require('../../../lib/jobs');
@@ -8,79 +7,7 @@ var JobsModule = require('../../../lib/jobs');
 require('../../../lib/app');
 
 describe('Jobs RESTFUL', function () {
-  var existingFile;
-  var uploadsPath = path.join(__dirname, "../../../uploads");
-  this.timeout(20000);
-
-  // recode existing files in uploads/configFile.json 
-  before(function (done) {
-    fs.readdir(uploadsPath, function (err, fileList) {
-      existingFile = fileList;
-      done();
-    });
-  });
-
-  // remove created test uploaded files and revert db changes
-  after(function (done) {
-    //remove uploaded files
-    fs.readdir(uploadsPath, function (err, fileList) {
-      fileList.forEach(function (file) {
-        if (existingFile.indexOf(file) == -1) {
-          fs.unlink(uploadsPath + '/' + file, function (err) {
-            if (err) {
-              throw err;
-            }
-          });
-        }
-      });
-      //remove db changes
-      db.collection('jobs').remove(
-        {$or: [
-          {name: "test job"},
-          {name: "db save job test"}
-          // {name: "job with id"}
-        ]},
-        function (err) {
-          if (err) {
-            throw err;
-          }
-          done();
-        });
-    });
-  });
-
-  it('should return 400 when name field is not legal', function(done) {
-    var ropts = {
-      url: 'http://localhost:3000/jobs',
-      headers: {
-        'Accept': 'application/json'
-      }
-    };
-    var r = request.post(ropts, callback);
-    var form = r.form();
-    form.append('name', '');
-    form.append('configFile', fs.createReadStream(__dirname + '/configFile.json'));
-
-    function callback(err, response) {
-      response.statusCode.should.equal(400);
-      done();
-    }    
-
-  });
-
-  it('should return 200 and JSON when post to /jobs ', function (done) {
-    var r = request.post('http://localhost:3000/jobs', callback);
-    var form = r.form();
-    form.append('name', 'test job');
-    form.append('configFile', fs.createReadStream(__dirname + '/configFile.json'));
-
-    function callback(err, response, body) {
-      var job = JSON.parse(body);
-      job.should.have.property('name','test job');
-      done();
-    }
-  });
-
+  
   it('should put attached file in %PROJECT_PATH/uploads from post /job', function (done) {
     var r = request.post('http://localhost:3000/jobs', callback);
     var form = r.form();
